Website/APIs_Docker/project_gpt/requirements.txt
flask
flask-cors
python-dotenv
openai
Pillow